03_mini_info_diet/requirements.txt
# Gmail API authentication and client
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
google-genai
# HTTP requests for OpenRouter API
requests>=2.28.0

# Environment variable management
python-dotenv>=1.0.0

PyMuPDF>=1.24.10
openreview-py>=1.0.0